image_recognition_demo/requirements.txt
huaweicloudsdkcore==3.1.70
huaweicloudsdkimage==3.1.70
requests==2.31.0
python-dotenv==1.0.0